+++ packages/candid/src/idl.test.ts
@@ -1112,6 +1112,12 @@ describe('IDL subtyping', () => {
     testSub(IDL.Tuple(IDL.Tuple(Odd)), Odd);
   });
 
+  describe('Subtyping on records/variants normalizes field labels', () => {
+    // Checks we don't regress https://github.com/dfinity/agent-js/issues/1072
+    testSub(IDL.Record({ a: IDL.Nat, "_98_": IDL.Nat }), IDL.Record({ "_97_": IDL.Nat, b: IDL.Nat }));
+    testSub(IDL.Variant({ a: IDL.Nat, "_98_": IDL.Nat }), IDL.Variant({ "_97_": IDL.Nat, b: IDL.Nat }));
+  });
+
   describe('decoding function/service references', () => {
     const principal = Principal.fromText('w7x7r-cok77-xa');
     it('checks subtyping when decoding function references', () => {